✓ samplingrate UINT16 -> UINT32
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui@2460 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ support_gimic/pccore.h
@@ -101,7 +101,7 @@ struct tagNP2Config
 	UINT32	BG_COLOR;
 	UINT32	FG_COLOR;
 
-	UINT16	samplingrate;
+	UINT32	samplingrate;
 	UINT16	delayms;
 	UINT8	SOUND_SW;
 	UINT8	snd_x;
